Add scalar field
