+++ algebra/elliptic-curve/src/lib.rs
@@ -45,6 +45,7 @@
 
 mod curve;
 mod jacobian;
+mod scalar_field;
 mod wnaf;
 
 #[cfg(not(feature = "std"))]
@@ -52,11 +53,12 @@ extern crate no_std_compat as std;
 
 pub use curve::Affine;
 pub use jacobian::Jacobian;
+pub use scalar_field::{Element as ScalarFieldElement, Order};
 pub use wnaf::{base_mul, double_base_mul, double_mul, mul, window_table_affine};
 
 use zkp_macros_decl::u256h;
 use zkp_primefield::FieldElement;
-use zkp_u256::{MontgomeryParameters, U256};
+use zkp_u256::U256;
 
 // Curve parameters
 
@@ -69,20 +71,6 @@ pub const BETA: FieldElement = FieldElement::from_montgomery_const(u256h!(
 ));
 
 pub const ORDER: U256 = u256h!("0800000000000010ffffffffffffffffb781126dcae7b2321e66a241adc64d2f");
-
-#[cfg_attr(feature = "std", derive(Debug))]
-pub struct Order();
-
-impl MontgomeryParameters for Order {
-    type UInt = U256;
-
-    const M64: u64 = 0xbb6b_3c4c_e8bd_e631;
-    const MODULUS: U256 =
-        u256h!("0800000000000010ffffffffffffffffb781126dcae7b2321e66a241adc64d2f");
-    const R1: U256 = u256h!("07fffffffffffdf10000000000000008c75ec4b46df16bee51925a0bf4fca74f");
-    const R2: U256 = u256h!("07d9e57c2333766ebaf0ab4cf78bbabb509cf64d14ce60b96021b3f1ea1c688d");
-    const R3: U256 = u256h!("01b2ba88ca1fe18a1f0d9dedfedfda501da2136eb8b3f20e81147668fddd0429");
-}
 
 // x = 0x01ef15c18599971b7beced415a40f0c7deacfd9b0d1819e03d723d8bc943cfca
 // y = 0x005668060aa49730b7be4801df46ec62de53ecd11abe43a32873000c36e8dc1f
